sso: replace Shader binding point with _Shader
V2: formating improvement

To avoid NULL pointer check a default pipeline object is installed in _Shader when no
program is current

The spec say that UseProgram/UseShaderProgramEXT/ActiveProgramEXT got an higher
priority over the pipeline object. When default program is uninstall, the pipeline is
used if any was bound.

Note: A careful rename need to be done now...